requirements.txt
PyPDF2 
tabula-py 
pandas
jpype1
treelib
openpyxl
rich
python-dotenv